Address stopped workflow review feedback
