+++ crates/harness-server/src/handlers/operator_monitor/sampling.rs
@@ -1,0 +1,46 @@
+use super::{OPERATOR_ACTION_STATES, WORKFLOW_DEFINITION_IDS, WORKFLOW_SAMPLE_LIMIT};
+use harness_workflow::runtime::{WorkflowInstance, WorkflowRuntimeStore};
+use std::cmp::Reverse;
+use std::collections::HashSet;
+
+pub(super) async fn list_operator_action_workflows(
+    store: &WorkflowRuntimeStore,
+) -> anyhow::Result<Vec<WorkflowInstance>> {
+    let mut workflows = Vec::new();
+    for definition_id in WORKFLOW_DEFINITION_IDS {
+        for state in OPERATOR_ACTION_STATES {
+            workflows.extend(
+                store
+                    .list_recent_instances_by_state(definition_id, state, WORKFLOW_SAMPLE_LIMIT)
+                    .await?,
+            );
+        }
+    }
+    Ok(workflows)
+}
+
+pub(super) fn dedupe_workflows(workflows: &mut Vec<WorkflowInstance>) {
+    let mut seen = HashSet::new();
+    workflows.retain(|workflow| seen.insert(workflow.id.clone()));
+}
+
+pub(super) async fn list_recent_failed_workflows(
+    store: &WorkflowRuntimeStore,
+    capacity: usize,
+) -> anyhow::Result<Vec<WorkflowInstance>> {
+    if capacity == 0 {
+        return Ok(Vec::new());
+    }
+    let per_definition_limit = capacity.min(WORKFLOW_SAMPLE_LIMIT as usize) as i64;
+    let mut workflows = Vec::new();
+    for definition_id in WORKFLOW_DEFINITION_IDS {
+        workflows.extend(
+            store
+                .list_recent_instances_by_state(definition_id, "failed", per_definition_limit)
+                .await?,
+        );
+    }
+    workflows.sort_by_key(|workflow| Reverse(workflow.updated_at));
+    workflows.truncate(capacity);
+    Ok(workflows)
+}
