+++ crates/harness-server/src/runtime_projection.rs
@@ -1,15 +1,15 @@
+mod stopped_actions;
+
 use crate::task_runner::{
     SchedulerAuthorityState, TaskFailureKind, TaskId, TaskPhase, TaskSchedulerState, TaskStatus,
 };
 use harness_workflow::runtime::{
-    ActivityErrorKind, WorkflowCommand, WorkflowCommandType, WorkflowInstance,
-    WorkflowRuntimeStore, WorkflowTerminalState, GITHUB_ISSUE_PR_DEFINITION_ID,
-    LOCAL_REVIEW_ACTIVITY, PR_FEEDBACK_DEFINITION_ID, PR_FEEDBACK_INSPECT_ACTIVITY,
-    QUALITY_GATE_DEFINITION_ID,
+    WorkflowInstance, WorkflowTerminalState, QUALITY_GATE_DEFINITION_ID,
 };
 use serde::Serialize;
 use serde_json::Value;
-use std::collections::HashMap;
+
+pub(crate) use stopped_actions::stopped_action_eligibility_for_workflows;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) enum RuntimeActiveBucket {
@@ -271,247 +271,6 @@ fn structured_last_stop(data: &serde_json::Value) -> Option<Value> {
     data.get("last_stop")
         .filter(|value| value.is_object())
         .cloned()
-}
-
-pub(crate) async fn stopped_action_eligibility_for_workflows(
-    store: Option<&WorkflowRuntimeStore>,
-    workflows: &[WorkflowInstance],
-) -> anyhow::Result<HashMap<String, RuntimeStoppedActionEligibility>> {
-    let Some(store) = store else {
-        return Ok(HashMap::new());
-    };
-
-    let mut plans = Vec::new();
-    let mut runtime_job_ids = Vec::new();
-    for workflow in workflows {
-        let Some(plan) = stopped_action_plan(workflow) else {
-            continue;
-        };
-        if let Some(runtime_job_id) = plan.runtime_job_id.as_ref() {
-            runtime_job_ids.push(runtime_job_id.clone());
-        }
-        plans.push((workflow.id.clone(), plan));
-    }
-
-    let command_sources = store
-        .command_sources_for_runtime_jobs(&runtime_job_ids)
-        .await?;
-    let mut by_workflow = HashMap::new();
-    for (workflow_id, plan) in plans {
-        let allowed = match plan.runtime_job_id.as_ref() {
-            None => plan.legacy_fallback,
-            Some(runtime_job_id) => command_sources
-                .get(runtime_job_id)
-                .filter(|source| source.workflow_id == workflow_id)
-                .is_some_and(|source| {
-                    command_matches_recovery_target(&source.command, plan.target)
-                }),
-        };
-        if allowed {
-            by_workflow.insert(
-                workflow_id,
-                match plan.action {
-                    StoppedRecoveryAction::Unblock => RuntimeStoppedActionEligibility {
-                        can_unblock: true,
-                        can_retry: false,
-                    },
-                    StoppedRecoveryAction::Retry => RuntimeStoppedActionEligibility {
-                        can_unblock: false,
-                        can_retry: true,
-                    },
-                },
-            );
-        }
-    }
-    Ok(by_workflow)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct StoppedActionPlan {
-    action: StoppedRecoveryAction,
-    target: RecoveryDispatchTarget,
-    runtime_job_id: Option<String>,
-    legacy_fallback: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum StoppedRecoveryAction {
-    Unblock,
-    Retry,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct RecoveryDispatchTarget {
-    activity: &'static str,
-}
-
-fn stopped_action_plan(workflow: &WorkflowInstance) -> Option<StoppedActionPlan> {
-    if workflow.definition_id != GITHUB_ISSUE_PR_DEFINITION_ID {
-        return None;
-    }
-    let action = match workflow.state.as_str() {
-        "blocked" => StoppedRecoveryAction::Unblock,
-        "failed" => StoppedRecoveryAction::Retry,
-        _ => return None,
-    };
-    let metadata = StoppedRecoveryMetadata::from_workflow_data(&workflow.data)?;
-    if action == StoppedRecoveryAction::Retry
-        && matches!(
-            metadata.error_kind,
-            Some(ActivityErrorKind::Fatal | ActivityErrorKind::Configuration)
-        )
-    {
-        return None;
-    }
-
-    let (target, runtime_job_id, legacy_fallback) = match metadata.activity.as_deref() {
-        Some(activity) => (
-            recovery_dispatch_target(activity)?,
-            Some(metadata.runtime_job_id?),
-            false,
-        ),
-        None if metadata.has_no_structured_stop_metadata => (
-            RecoveryDispatchTarget {
-                activity: "implement_issue",
-            },
-            None,
-            true,
-        ),
-        None => return None,
-    };
-    Some(StoppedActionPlan {
-        action,
-        target,
-        runtime_job_id,
-        legacy_fallback,
-    })
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct StoppedRecoveryMetadata {
-    activity: Option<String>,
-    runtime_job_id: Option<String>,
-    error_kind: Option<ActivityErrorKind>,
-    has_no_structured_stop_metadata: bool,
-}
-
-impl StoppedRecoveryMetadata {
-    fn from_workflow_data(data: &Value) -> Option<Self> {
-        if data
-            .get("last_stop")
-            .filter(|value| !value.is_null())
-            .is_some_and(|value| !value.is_object())
-        {
-            return None;
-        }
-        let error_kind = stopped_error_kind(data)?;
-        let _state = optional_metadata_string(data.pointer("/last_stop/state"))?;
-        Some(Self {
-            activity: optional_metadata_string(data.pointer("/last_stop/activity"))?,
-            runtime_job_id: optional_metadata_string(data.pointer("/last_stop/runtime_job_id"))?,
-            has_no_structured_stop_metadata: data.get("last_stop").is_none_or(Value::is_null)
-                && error_kind.is_none(),
-            error_kind,
-        })
-    }
-}
-
-fn stopped_error_kind(data: &Value) -> Option<Option<ActivityErrorKind>> {
-    let root = optional_error_kind(data.get("error_kind"))?;
-    let last_stop = optional_error_kind(data.pointer("/last_stop/error_kind"))?;
-    Some(root.or(last_stop))
-}
-
-fn optional_metadata_string(value: Option<&Value>) -> Option<Option<String>> {
-    let Some(value) = value.filter(|value| !value.is_null()) else {
-        return Some(None);
-    };
-    value
-        .as_str()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(|value| Some(value.to_string()))
-}
-
-fn optional_error_kind(value: Option<&Value>) -> Option<Option<ActivityErrorKind>> {
-    let Some(value) = value.filter(|value| !value.is_null()) else {
-        return Some(None);
-    };
-    if value.as_str().is_some_and(|value| value.trim().is_empty()) {
-        return None;
-    }
-    serde_json::from_value(value.clone()).ok().map(Some)
-}
-
-fn recovery_dispatch_target(activity: &str) -> Option<RecoveryDispatchTarget> {
-    match activity {
-        "implement_issue" => Some(RecoveryDispatchTarget {
-            activity: "implement_issue",
-        }),
-        "replan_issue" => Some(RecoveryDispatchTarget {
-            activity: "replan_issue",
-        }),
-        "merge_pr" => Some(RecoveryDispatchTarget {
-            activity: "merge_pr",
-        }),
-        LOCAL_REVIEW_ACTIVITY => Some(RecoveryDispatchTarget {
-            activity: LOCAL_REVIEW_ACTIVITY,
-        }),
-        "sweep_pr_feedback" => Some(RecoveryDispatchTarget {
-            activity: "sweep_pr_feedback",
-        }),
-        PR_FEEDBACK_INSPECT_ACTIVITY => Some(RecoveryDispatchTarget {
-            activity: PR_FEEDBACK_INSPECT_ACTIVITY,
-        }),
-        "start_child_workflow" => Some(RecoveryDispatchTarget {
-            activity: "start_child_workflow",
-        }),
-        "address_pr_feedback" => Some(RecoveryDispatchTarget {
-            activity: "address_pr_feedback",
-        }),
-        _ => None,
-    }
-}
-
-fn command_matches_recovery_target(
-    command: &WorkflowCommand,
-    target: RecoveryDispatchTarget,
-) -> bool {
-    match command.command_type {
-        WorkflowCommandType::EnqueueActivity => {
-            command.activity_name() == Some(target.activity)
-                && enqueue_payload_matches_target(&command.command)
-        }
-        WorkflowCommandType::StartChildWorkflow => {
-            let payload = &command.command;
-            matches!(
-                target.activity,
-                "start_child_workflow" | "sweep_pr_feedback"
-            ) && payload.get("definition_id").and_then(Value::as_str)
-                == Some(PR_FEEDBACK_DEFINITION_ID)
-                && payload.get("child_activity").and_then(Value::as_str)
-                    == Some(PR_FEEDBACK_INSPECT_ACTIVITY)
-                && payload.get("pr_number").and_then(Value::as_u64).is_some()
-                && payload
-                    .get("subject_key")
-                    .and_then(Value::as_str)
-                    .is_some_and(|value| !value.trim().is_empty())
-        }
-        _ => false,
-    }
-}
-
-fn enqueue_payload_matches_target(payload: &Value) -> bool {
-    let review_summary = payload
-        .get("review_summary")
-        .and_then(Value::as_str)
-        .is_some_and(|value| !value.trim().is_empty());
-    let hygiene = payload
-        .get("hygiene")
-        .or_else(|| payload.get("hygiene_context"))
-        .is_some_and(|value| !value.is_null());
-    payload.get("source").and_then(Value::as_str) != Some("pr_hygiene")
-        || (payload.get("pr_number").and_then(Value::as_u64).is_some() && review_summary && hygiene)
 }
 
 fn legacy_dedupe_task_handle(data: &serde_json::Value) -> Option<TaskId> {
